backend/requirements.txt
fastapi==0.116.1
uvicorn==0.33.0
requests==2.32.4
pillow==10.2.0
numpy==1.24.1
diffusers==0.35.1
opencv-python==4.12.0.88
transformers==4.46.3
accelerate==1.0.1
python-multipart==0.0.20